From: Seunggyun Kim <sgyun.kim@samsung.com> Subject: [E-devel] [Patch] elm_genlist - added auto scroll function when expanding items.
Now, The user don't know how many items are added when the item is expanded in tree view.
If the use want to know how many items are added, they have to scroll
genlist view.

So I added auto scroll function when the items is expanded on bottom of viewport.
Using this, the user can know expanded items although that is expanded
beyond the view port.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@60302 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33