elementary/layout - added theme changed hook
Because the user can use their' own edj objs and call the signals for the edj objs,
They have a responsibility to reset the status of the edj objs when layout theme is changed.

This "theme,changed" signal provides to user a chance to reset (signal-emit) again.

Author: hermet
Date: 2011-07-20 01:36:40 -0700 (Wed, 20 Jul 2011)
New Revision: 61518
Trac: http://trac.enlightenment.org/e/changeset/61518