[accessability] Initial merge
this is... the beginning of accessibility supportin elm. it's direct
as in elm manages it itself - all it needs is a module to send text
to. one is provided here that just execs espeak and handles a stream
of things for it to say. this is only a start and is still being
fleshed out.

git-svn-id: https://svn.enlightenment.org/svn/e/trunk/elementary@62585 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33