new toolbar mode: always_select
when this mode is enabled, a toolbar item will always be selected. if the selected item is deleted, the next item will be selected. if there is no next item, the first item will be selected. if there is no next or first item, you can go <expletive deleted> yourself.

git-svn-id: https://svn.enlightenment.org/svn/e/trunk/elementary@63691 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33