From: Jaehwan Kim <jae.hwan.kim@samsung.com> Subject: [E-devel] [Patch] els_scroller - bouncing doesn't work in specific case
I found the bouncing of the scroller doesn't work in specific case.
When the scroller is scrolled but the animation doesn't occur,
sd->down.dy is not changed. Then in elm_smart_scroller_child_pos_set,
it's possible not to pass the below conditional sentence even if it
have to has the bounce.

if (((y < miny) && (0 <= sd->down.dy)) ||
((y > (my + miny)) && (0 >= sd->down.dy)))

So I reset the value sd->down.dx and sd->down.dy to 0 when mouse is up.
Please check patch file.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@63797 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33