From: Jaehwan Kim <jae.hwan.kim@samsung.com> Subject: [E-devel] [Patch] ELM_THUMBSCROLL_SENSITIVITY_FRICTION
I talked with you about 4.0, magic number.
I want to change the part that the length of mouse dragging multiply 4.0
to _elm_config->thumbscroll_sensitivity_friction.
4.0 is nice. But some device may want to change the value. It is a kind
of sensitivity. So I added the config "ELM_THUMBSCROLL_SENESITIVITY
_FRICTION". And I change the value mean. If the value is high, it is
more sensitive. The range of the value is from 0.1 to 1.0. It cannot be 0.0.
4.0 changed to 0.25. 
Please check this patch.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@64051 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33