[*] merge with svn revision 64406
From: Jaehwan Kim <jae.hwan.kim@samsung.com>

I found a problem in els_scroller. When the scroller is resized, it shows the location of
the coordination which was saved into wanted_coordination. If you don't use the page,
it is no matter. But if you use the page function, the animation of the scroller is operated
by _smart_scrollto_x_animator, not by _smart_momentum_animator.
In _smart_scrollto_x_animator, therer isn't _update_wanted_coordinates. So the value of
wanted_coordination doesn't update. I think _update_wanted_coordinates have to be added in
_smart_scrollto_x_animator and _smart_scrollto_y_animator.
Please check the attached patch file.

Change-Id: I14f92c38047d165f5df322e1b25637c1794a0044
git-svn-id: https://svn.enlightenment.org/svn/e/trunk/elementary@64406 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33