elementary/scroller - change the behavior when the scroller is bouncing.
When the scroller is dragged repeatedly in a short time, it is accelerated.
If it reaches the edge of the content, the bouncing is started.
Sometimes because of the acceleration, the content cannot be shown in a moment.
In order to avoid that situation, the bounce animation have to be fixed 
according the amount of acceleration.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@64642 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33