From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [Patch][elementary] elm_list.c - item_content_set() related.
The Eina_Bool item->dummy_icon and item->dummy_end of Elm_List_Item
indicate whether the content has REAL object or not.
- "not" means.. if elm_object_item_part_content_set(); is called with
NULL content, the Elm_List set the swallow part with
rectangle which has alpha value 0 and the item->dummy_icon or
item->dummy_end is set to EINA_TRUE. -
But in the item_content_set(); it works the opposite way. So when
content set with NULL value the flag should be EINA_TRUE.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@69535 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33