From: GoUn Lee <gouni.lee@samsung.com> Subject: [E-devel] [?Patch]elm_?multibutto?nentry bug fix
I've attached patch to fix bug in multibuttonentry.
When this widget is unfocused, there's no code for making unfocused
entry in multibuttonentry.
Because of that, when multibuttonentry got focus again, entry couldn't
get focus normally.
so I've added  code to fix this issue. when it's unfocused.

Also I removed unused value in resize_button() which is 'padding_inner'
 


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@69583 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33