From: cnook <kimcinoo@gmail.com> Subject: [E-devel] [Patch][elementary_test] test_list
I had added test code for the elementary. but I made a pause it
because of
some reasons and just started again.
Today I have attached a patch for the elm_list which does not have test
code for the API elm_list_item_prev();
Please review the attached patch and give any feedbacks. I'll be back
with
other patches. Thanks!



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/elementary@70265 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33